// src/components/Dashboard/BeverageData/Label/utils/FieldName.enum.ts
enum FieldName {
  badge = 'badge',
  // -----------
  name = 'name',
  series = 'series',
  brand = 'brand',
  cooperation = 'cooperation',
  contract = 'contract',
  place = 'place',
  tale = 'tale',
  barcode = 'barcode',
  // -----------
  fermentation = 'fermentation',
  style = 'style',
  extract = 'extract',
  alcohol = 'alcohol',
  filtration = 'filtration',
  pasteurization = 'pasteurization',
  aged = 'aged',
  dryHopped = 'dryHopped',
  expirationDate = 'expirationDate',
  // -----------
  ingredients = 'ingredients',
  ingredientsList = 'ingredientsList',
  smokedMalt = 'smokedMalt',
  // -----------
  bitterness = 'bitterness',
  sweetness = 'sweetness',
  fullness = 'fullness',
  power = 'power',
  hoppyness = 'hoppyness',
  temperature = 'temperature',
  // -----------
  container = 'container',
  price = 'price',
}

export default FieldName;